requirements.txt
fastapi
uvicorn
python-dotenv
groq
pydantic
google-images-search
